chore: refactoring logic back inside action menu
